+++ applications/fishing-map/src/features/datasets/datasets.hook.ts
@@ -117,7 +117,7 @@ export const useDatasetsAPI = () => {
       if (fetchDatasetByIdThunk.fulfilled.match(action)) {
         return { payload: action.payload }
       } else {
-        return { error: action.payload }
+        return { error: action.payload as AsyncError }
       }
     },
     [dispatch]
@@ -129,7 +129,7 @@ export const useDatasetsAPI = () => {
       if (createDatasetThunk.fulfilled.match(action)) {
         return { payload: action.payload }
       } else {
-        return { error: action.payload }
+        return { error: action.payload as AsyncError }
       }
     },
     [dispatch]
@@ -143,7 +143,7 @@ export const useDatasetsAPI = () => {
       if (updateDatasetThunk.fulfilled.match(action)) {
         return { payload: action.payload }
       } else {
-        return { error: action.payload }
+        return { error: action.payload as AsyncError }
       }
     },
     [dispatch]
@@ -177,7 +177,7 @@ export const useCarrierLatestConnect = () => {
     if (fetchLastestCarrierDatasetThunk.fulfilled.match(action)) {
       return { payload: action.payload }
     } else {
-      return { error: action.payload }
+      return { error: action.payload as AsyncError }
     }
   }, [dispatch])