+++ applications/fishing-map/src/features/map/map-features.hooks.ts
@@ -92,14 +92,16 @@ export const useHasSourceLoaded = (sourceId: string) => {
 export const useActiveHeatmapAnimatedMetadatas = (generators: AnyGeneratorConfig[]) => {
   const style = useMapStyle()
   const generatorsIds = generators.map((generator) => generator.id)
-  const generatorsMetadata = generatorsIds.map((generatorId) => {
-    return style?.metadata?.generatorsMetadata[generatorId]
+  const generatorsMetadata = generatorsIds.flatMap((generatorId) => {
+    return style?.metadata?.generatorsMetadata[generatorId] || []
   })
+
   const serializedGeneratorIds = generatorsMetadata
     .map((metadata) => {
-      return metadata?.timeChunks?.activeSourceId
+      return metadata?.timeChunks?.activeSourceId + metadata?.numSublayers
     })
     .join()
+
   const metadatas = useMemo(() => {
     return generatorsMetadata
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -137,6 +139,7 @@ export const useFeatures = ({
 
 const useGeneratorAnimatedFeatures = (generators: AnyGeneratorConfig[]) => {
   const sourcesMetadata = useActiveHeatmapAnimatedMetadatas(generators)
+
   const sourcesIds: string[] = useMemo(() => {
     return sourcesMetadata.map((metadata) => metadata?.timeChunks?.activeSourceId)
   }, [sourcesMetadata])